jest: Add Jest ESM example
