+++ examples/protobufjs/readme.md
@@ -1,0 +1,5 @@
+# protobufjs fuzz test
+
+This fuzz test uses
+[ECMAScript Modules](https://jestjs.io/docs/ecmascript-modules) and requires
+Node.js 18.8.0+.
